created cli directory for command implementation
